+++ eras/mary/impl/src/Cardano/Ledger/Mary/BlockBody.hs
@@ -2,7 +2,9 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -10,8 +12,19 @@
 module Cardano.Ledger.Mary.BlockBody () where
 
 import Cardano.Ledger.BaseTypes (ProtVer (..))
-import Cardano.Ledger.Binary (EncCBORGroup (..), serialize')
-import Cardano.Ledger.Core (EraBlockBody (..))
+import Cardano.Ledger.Binary (
+  Annotator,
+  DecCBOR (decCBOR),
+  EncCBOR (..),
+  EncCBORGroup (..),
+  decodeRecordNamed,
+  encodeListLen,
+  serialize',
+  toPlainEncoding,
+ )
+import qualified Cardano.Ledger.Binary.Plain as Plain
+import Cardano.Ledger.Block (Block (..))
+import Cardano.Ledger.Core (EraBlockBody (..), eraProtVerLow)
 import Cardano.Ledger.Mary.Era (MaryEra)
 import Cardano.Ledger.Mary.Tx ()
 import Cardano.Ledger.Shelley.BlockBody (
@@ -21,11 +34,28 @@ import Cardano.Ledger.Shelley.BlockBody (
   txSeqBlockBodyShelleyL,
  )
 import qualified Data.ByteString as BS
+import Data.Typeable (Typeable)
 
 instance EraBlockBody MaryEra where
   type BlockBody MaryEra = ShelleyBlockBody MaryEra
   mkBasicBlockBody = mkBasicBlockBodyShelley
   txSeqBlockBodyL = txSeqBlockBodyShelleyL
   hashBlockBody = shelleyBlockBodyHash
-  numSegComponents = 3
   blockBodySize (ProtVer v _) = BS.length . serialize' v . encCBORGroup
+
+instance EncCBOR h => EncCBOR (Block h MaryEra) where
+  encCBOR (Block h txns) =
+    encodeListLen 4 <> encCBOR h <> encCBORGroup txns
+
+instance (EncCBOR h, Typeable h) => Plain.ToCBOR (Block h MaryEra) where
+  toCBOR = toPlainEncoding (eraProtVerLow @MaryEra) . encCBOR
+
+instance
+  (DecCBOR (Annotator h), Typeable h) =>
+  DecCBOR (Annotator (Block h MaryEra))
+  where
+  decCBOR =
+    decodeRecordNamed "Block" (const 4) $ do
+      header <- decCBOR
+      txns <- decCBOR
+      pure $ Block <$> header <*> txns