+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Binary/Annotator.hs
@@ -15,24 +15,11 @@ module Test.Cardano.Ledger.Core.Binary.Annotator (
 ) where
 
 import Cardano.Ledger.Binary
-import Cardano.Ledger.Block
 import Cardano.Ledger.Core
 import Cardano.Ledger.Plutus
 import Test.Cardano.Ledger.Binary (decoderEquivalenceSpec)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Arbitrary ()
-
-instance
-  ( EraBlockBody era
-  , DecCBOR h
-  , DecCBOR (BlockBody era)
-  ) =>
-  DecCBOR (Block h era)
-  where
-  decCBOR =
-    decodeRecordNamed "Block" (const blockSize) $ Block <$> decCBOR <*> decCBOR
-    where
-      blockSize = 1 + fromIntegral (numSegComponents @era)
 
 decoderEquivalenceEraSpec ::
   forall era t.